+++ src/BadgeSelectItemBuilder/test/BadgeSelectItemBuilder.tsx
@@ -1,7 +1,8 @@
 import { badgeSelectItemBuilder } from '..';
+import * as React from 'react';
 
 function badgeSelectItemBuilderWithAllProps() {
-  const { id, value } = badgeSelectItemBuilder({
+  const { id, value , overrideStyle} = badgeSelectItemBuilder({
     id: '1',
     skin: 'danger',
     text: 'text',